add task reference router
refactor: plan searching middleware, subtasks naming
