+++ miamoremustachio/ToDo/modules/models/task-model.js
@@ -5,7 +5,7 @@ const { TO_DO } = STATUSES;
 const { LOW } = PRIORITIES;
 const { getDefaultDeadline } = require('../helpers/task-helper.js');
 
-const subtasksSchema = new mongoose.Schema({
+const subtaskSchema = new mongoose.Schema({
   title: String,
   completed: { type: Boolean, default: false },
 });
@@ -15,7 +15,7 @@ const taskSchema = new mongoose.Schema({
   status: { type: String, default: TO_DO },
   priority: { type: String, default: LOW },
   deadline: { type: Date, default: getDefaultDeadline },
-  subtasks: [subtasksSchema],
+  subtasks: [subtaskSchema],
   user: { type: 'ObjectId', ref: 'User' },
 },
 {
