+++ miamoremustachio/ToDo/modules/routes/plan-routes.js
@@ -72,57 +72,4 @@ router.route('/:id')
     }
   });
 
-router.route('/:id/tasks')
-  .all(findPlan, checkUserId)
-  .post(async (req, res) => {
-    const planId = req.params.id;
-
-    try {
-      const plan = await plans.get(planId);
-      const tasksList = plan.tasks;
-
-      plans.addTask(tasksList, req.body);
-
-      await plans.save(plan);
-
-      const taskPath = `${req.originalUrl}/${planId}`;
-
-      res.send(taskPath);
-    } catch(error) {
-      res.send(500).send(error.message);
-    }
-  });
-
-router.route('/:id/tasks/:taskId')
-  .all(findPlan, checkUserId)
-  .get(async (req, res) => {
-    const planId = req.params.id;
-    const taskId = req.params.taskId;
-
-    try {
-      const plan = await plans.get(planId);
-      const tasksList = plan.tasks;
-      const task = plans.getTask(tasksList, taskId);
-
-      res.json(task);
-    } catch(error) {
-      res.status(500).send(error.message);
-    }
-  })
-  .delete(async (req, res) => {
-    const planId = req.params.id;
-    const taskId = req.params.taskId;
-
-    try {
-      const plan = await plans.get(planId);
-      plan.tasks = plans.deleteTask(plan.tasks, taskId);
-      
-      await plans.save(plan);
-
-      res.sendStatus(204);
-    } catch(error) {
-      res.status(500).send(error.message);
-    }
-  });
-
 module.exports = { router };
